// src/types/TenancyType.tsx
{/*  ===================================THIS FILE WAS AUTO GENERATED=================================== */}

export type TenancyType = {
    tenancyTypeID: number;
    code: string;
    name: string;
    description: string;
    minimumMonths: number;
    maximumMonths: number;
    isPeriodic: boolean;
    isActive: boolean;
    capturedDate: string;
    capturedBy: string;
};